+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/ChannelInfoMapping.kt
@@ -10,4 +10,5 @@ internal fun ChannelInfoDto.toDomain(): ChannelInfo =
         memberCount = member_count,
         name = name,
         type = type,
+        image = image,
     )