+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/UserDtos.kt
@@ -11,6 +11,8 @@ import java.util.Date
 internal data class UpstreamUserDto(
     val banned: Boolean,
     val id: String,
+    val name: String,
+    val image: String,
     val invisible: Boolean,
     val role: String,
     val devices: List<DeviceDto>,
@@ -26,6 +28,8 @@ internal data class UpstreamUserDto(
 @JsonClass(generateAdapter = true)
 internal data class DownstreamUserDto(
     val id: String,
+    val name: String = "",
+    val image: String = "",
     val role: String,
     val invisible: Boolean = false,
     val banned: Boolean,